+++ nb/CliParm.cpp
@@ -94,7 +94,7 @@ CliParm* CliParm::AccessParm(CliCookie& cookie, uint32_t depth) const
 
 ptrdiff_t CliParm::CellDiff()
 {
-   int local;
+   uintptr_t local;
    auto fake = reinterpret_cast< const CliParm* >(&local);
    return ptrdiff(&fake->pid_, fake);
 }
